refactor hex/cnv parsing
